+++ src/main/java/com/vca/repositories/SegmentRepository.java
@@ -3,6 +3,6 @@ package com.vca.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.vca.entity.Segment;
 
-public interface SegmentRepository extends JpaRepository<Segment, Long> {
+public interface SegmentRepository extends JpaRepository<Segment, Integer> {
 
 }
